Add codeowners Terraform Data Source
This data source provides an existence check for a CODEOWNERS file
within the specified repository.

The intent of this is to use as a toggle for enforcing codeowner reviews
in a repository resource. If codeowners are enforced, with no CODEOWNERS
file present, only repository admins will be able to accept pull
requests.

The following locations will be checked (as per Github docs):
- CODEOWNERS
- .github/CODEOWNERS
- docs/CODEOWNERS

REF: https://help.github.com/en/github/creating-cloning-and-archiving-re
positories/about-code-owners#codeowners-file-location