Remove email property from user queries
This property is inaccessible by some tokens - namely Github App
server-to-server tokens. Whereas REST queries simply return a null
value, GraphQL queries result in a "resource not accessible by
integration" error.

This property is not necessary in the context of Terraform management
of Github resources anyway, compatibility is of greater concern.